# ♻️ Refactor: reduce the memory usage of RoutePatternMatch