@ksw2000 2025-03-01
> `155-185`: **Consider consistent usage of byte slices or strings to reduce repeated conversions.**
> 
> You've successfully switched to byte slices for `patternPretty` but frequently convert it back to a string (lines 167, 169, 172, 185). This repeated conversion could introduce unnecessary overhead. As a good-to-have improvement, consider either consistently using byte slices throughout the matching logic or, if strings are ultimately needed, maintaining a single string variable for comparisons.

In lines 169, 172, and 185, the program does not allocate memory when converting between strings and byte slices. In line 167, I am still working on reducing the memory usage of `parseRoute`.