@gaby 2025-03-01
@ksw2000 I actually ran into the same problem with `path.go`. It wouldn't consistently reduce memory, so I never pushed the code 😂