@ksw2000 2025-03-01
> @ksw2000 I actually ran into the same problem with `path.go`. It wouldn't consistently reduce memory, so I never pushed the code 😂

I tried for the whole day, and I think this is a big issue that cannot be solved in one go. It needs to be optimized in multiple steps. Some functions can be refactored to be more concise. For example, `findLastCharsetPosition` and `findNextCharsetPosition` can return their values earlier.